// as/assembly/gen/tri.ts
export class Tri_default_hz_scalar_offset_audio_trig_audio {
  static readonly defaultInstance: Tri_default_hz_scalar_offset_audio_trig_audio = new Tri_default_hz_scalar_offset_audio_trig_audio()

  angle: f32 = 0
  lastOutput: f32 = 0
  lastHz: f32 = Infinity
  lastSampleRate: f32 = Infinity
  prevTrig: f32
  phaseInc: f32

  reset(): void {
    this.copyFrom(Tri_default_hz_scalar_offset_audio_trig_audio.defaultInstance)
  }

  copyFrom(src: Tri_default_hz_scalar_offset_audio_trig_audio): void {
    this.angle = src.angle
    this.lastOutput = src.lastOutput
    this.lastHz = src.lastHz
    this.lastSampleRate = src.lastSampleRate
    this.prevTrig = src.prevTrig
    this.phaseInc = src.phaseInc
  }

  process(bufferLength: i32, sampleCount: i32, sampleRate: f32, nyquist: f32, piOverNyquist: f32, bpm: f32, co: f32, samplesPerBeat: f32, samplesPerBar: f32, input$: usize, output$: usize, hz: f32, offset$: usize, trig$: usize): void {
    const hzClamped: f32 = max(hz, 0)
    const hzChanged: boolean = hzClamped !== this.lastHz
    const sampleRateChanged: boolean = sampleRate !== this.lastSampleRate


    if (hzChanged || sampleRateChanged) {
      if ((hzClamped > 0)) {
        this.phaseInc = (hzClamped / sampleRate)
      }
      this.lastHz = hzClamped
      this.lastSampleRate = sampleRate
    }

    let angle: f32 = this.angle
    let lastOutput: f32 = this.lastOutput
    let output: f32
    let phaseInc: f32 = this.phaseInc

    let integrated: f32
    let saw: f32

    let trig: f32

    let prevTrig: f32 = this.prevTrig
    for (let i = 0; i < bufferLength; i += 16) {
      unroll(16, () => {
        trig = load<f32>(trig$)
        if (((trig > 0) && (prevTrig <= 0))) {
          const offsetClamped: f32 = clamp(load<f32>(offset$), 0, 1)
          angle = offsetClamped
          lastOutput = 0
        }
        prevTrig = trig
        saw = (((2 * angle) - 1) - polyBlep(angle, phaseInc))
        integrated = ((phaseInc * saw) + ((1 - phaseInc) * lastOutput))
        output = (integrated * 6)
        lastOutput = integrated
        angle = (angle + phaseInc)
        if ((angle >= 1)) {
          angle = fract01(angle)
        }
        store<f32>(output$, output)
        output$ += 4
        offset$ += 4
        trig$ += 4
      })
    }

    this.angle = angle
    this.lastOutput = lastOutput
    this.prevTrig = prevTrig
  }
}

export class Tri_default_hz_audio_offset_scalar_trig_scalar {
  static readonly defaultInstance: Tri_default_hz_audio_offset_scalar_trig_scalar = new Tri_default_hz_audio_offset_scalar_trig_scalar()

  angle: f32 = 0
  lastOutput: f32 = 0
  lastTrig: f32 = Infinity
  prevTrig: f32
  phaseInc: f32

  reset(): void {
    this.copyFrom(Tri_default_hz_audio_offset_scalar_trig_scalar.defaultInstance)
  }

  copyFrom(src: Tri_default_hz_audio_offset_scalar_trig_scalar): void {
    this.angle = src.angle
    this.lastOutput = src.lastOutput
    this.lastTrig = src.lastTrig
    this.prevTrig = src.prevTrig
    this.phaseInc = src.phaseInc
  }

  process(bufferLength: i32, sampleCount: i32, sampleRate: f32, nyquist: f32, piOverNyquist: f32, bpm: f32, co: f32, samplesPerBeat: f32, samplesPerBar: f32, input$: usize, output$: usize, offset: f32, trig: f32, hz$: usize): void {
    const offsetClamped: f32 = clamp(offset, 0, 1)
    const trigChanged: boolean = trig !== this.lastTrig


    if (trigChanged) {
      if (((trig > 0) && (this.prevTrig <= 0))) {
        this.angle = offsetClamped
        this.lastOutput = 0
      }
      this.prevTrig = trig
      this.lastTrig = trig
    }

    let angle: f32 = this.angle
    let lastOutput: f32 = this.lastOutput
    let output: f32
    let phaseInc: f32 = this.phaseInc

    let integrated: f32
    let saw: f32

    let hzClamped: f32
    for (let i = 0; i < bufferLength; i += 16) {
      unroll(16, () => {
        hzClamped = max(load<f32>(hz$), 0)
        if ((hzClamped > 0)) {
          phaseInc = (hzClamped / sampleRate)
        }
        saw = (((2 * angle) - 1) - polyBlep(angle, phaseInc))
        integrated = ((phaseInc * saw) + ((1 - phaseInc) * lastOutput))
        output = (integrated * 6)
        lastOutput = integrated
        angle = (angle + phaseInc)
        if ((angle >= 1)) {
          angle = fract01(angle)
        }
        store<f32>(output$, output)
        output$ += 4
        hz$ += 4
      })
    }

    this.angle = angle
    this.lastOutput = lastOutput
  }
}

export class Tri_default_hz_audio_offset_scalar_trig_audio {
  static readonly defaultInstance: Tri_default_hz_audio_offset_scalar_trig_audio = new Tri_default_hz_audio_offset_scalar_trig_audio()

  angle: f32 = 0
  lastOutput: f32 = 0
  prevTrig: f32
  phaseInc: f32

  reset(): void {
    this.copyFrom(Tri_default_hz_audio_offset_scalar_trig_audio.defaultInstance)
  }

  copyFrom(src: Tri_default_hz_audio_offset_scalar_trig_audio): void {
    this.angle = src.angle
    this.lastOutput = src.lastOutput
    this.prevTrig = src.prevTrig
    this.phaseInc = src.phaseInc
  }

  process(bufferLength: i32, sampleCount: i32, sampleRate: f32, nyquist: f32, piOverNyquist: f32, bpm: f32, co: f32, samplesPerBeat: f32, samplesPerBar: f32, input$: usize, output$: usize, offset: f32, hz$: usize, trig$: usize): void {
    const offsetClamped: f32 = clamp(offset, 0, 1)

    let angle: f32 = this.angle
    let lastOutput: f32 = this.lastOutput
    let output: f32
    let phaseInc: f32 = this.phaseInc

    let integrated: f32
    let saw: f32

    let hzClamped: f32
    let trig: f32

    let prevTrig: f32 = this.prevTrig
    for (let i = 0; i < bufferLength; i += 16) {
      unroll(16, () => {
        hzClamped = max(load<f32>(hz$), 0)
        trig = load<f32>(trig$)
        if (((trig > 0) && (prevTrig <= 0))) {
          angle = offsetClamped
          lastOutput = 0
        }
        prevTrig = trig
        if ((hzClamped > 0)) {
          phaseInc = (hzClamped / sampleRate)
        }
        saw = (((2 * angle) - 1) - polyBlep(angle, phaseInc))
        integrated = ((phaseInc * saw) + ((1 - phaseInc) * lastOutput))
        output = (integrated * 6)
        lastOutput = integrated
        angle = (angle + phaseInc)
        if ((angle >= 1)) {
          angle = fract01(angle)
        }
        store<f32>(output$, output)
        output$ += 4
        hz$ += 4
        trig$ += 4
      })
    }

    this.angle = angle
    this.lastOutput = lastOutput
    this.prevTrig = prevTrig
  }
}

export class Tri_default_hz_audio_offset_audio_trig_scalar {
  static readonly defaultInstance: Tri_default_hz_audio_offset_audio_trig_scalar = new Tri_default_hz_audio_offset_audio_trig_scalar()

  angle: f32 = 0
  lastOutput: f32 = 0
  lastTrig: f32 = Infinity
  prevTrig: f32
  phaseInc: f32

  reset(): void {
    this.copyFrom(Tri_default_hz_audio_offset_audio_trig_scalar.defaultInstance)
  }

  copyFrom(src: Tri_default_hz_audio_offset_audio_trig_scalar): void {
    this.angle = src.angle
    this.lastOutput = src.lastOutput
    this.lastTrig = src.lastTrig
    this.prevTrig = src.prevTrig
    this.phaseInc = src.phaseInc
  }

  process(bufferLength: i32, sampleCount: i32, sampleRate: f32, nyquist: f32, piOverNyquist: f32, bpm: f32, co: f32, samplesPerBeat: f32, samplesPerBar: f32, input$: usize, output$: usize, trig: f32, hz$: usize, offset$: usize): void {
    const trigChanged: boolean = trig !== this.lastTrig


    if (trigChanged) {
      this.prevTrig = trig
      this.lastTrig = trig
    }

    let angle: f32 = this.angle
    let lastOutput: f32 = this.lastOutput
    let output: f32
    let phaseInc: f32 = this.phaseInc

    let integrated: f32
    let saw: f32

    let hzClamped: f32
    for (let i = 0; i < bufferLength; i += 16) {
      unroll(16, () => {
        hzClamped = max(load<f32>(hz$), 0)
        if ((hzClamped > 0)) {
          phaseInc = (hzClamped / sampleRate)
        }
        saw = (((2 * angle) - 1) - polyBlep(angle, phaseInc))
        integrated = ((phaseInc * saw) + ((1 - phaseInc) * lastOutput))
        output = (integrated * 6)
        lastOutput = integrated
        angle = (angle + phaseInc)
        if ((angle >= 1)) {
          angle = fract01(angle)
        }
        store<f32>(output$, output)
        output$ += 4
        hz$ += 4
        offset$ += 4
      })
    }

    this.angle = angle
    this.lastOutput = lastOutput
  }
}

export class Tri_default_hz_audio_offset_audio_trig_audio {
  static readonly defaultInstance: Tri_default_hz_audio_offset_audio_trig_audio = new Tri_default_hz_audio_offset_audio_trig_audio()

  angle: f32 = 0
  lastOutput: f32 = 0
  lastSampleRate: f32 = Infinity
  prevTrig: f32
  phaseInc: f32

  reset(): void {
    this.copyFrom(Tri_default_hz_audio_offset_audio_trig_audio.defaultInstance)
  }

  copyFrom(src: Tri_default_hz_audio_offset_audio_trig_audio): void {
    this.angle = src.angle
    this.lastOutput = src.lastOutput
    this.lastSampleRate = src.lastSampleRate
    this.prevTrig = src.prevTrig
    this.phaseInc = src.phaseInc
  }

  process(bufferLength: i32, sampleCount: i32, sampleRate: f32, nyquist: f32, piOverNyquist: f32, bpm: f32, co: f32, samplesPerBeat: f32, samplesPerBar: f32, input$: usize, output$: usize, hz$: usize, offset$: usize, trig$: usize): void {
    const sampleRateChanged: boolean = sampleRate !== this.lastSampleRate


    if (sampleRateChanged) {
      this.lastSampleRate = sampleRate
    }


    let angle: f32 = this.angle
    let lastOutput: f32 = this.lastOutput
    let output: f32
    let phaseInc: f32 = this.phaseInc

    let integrated: f32
    let saw: f32

    let hzClamped: f32
    let trig: f32

    let prevTrig: f32 = this.prevTrig
    for (let i = 0; i < bufferLength; i += 16) {
      unroll(16, () => {
        hzClamped = max(load<f32>(hz$), 0)
        trig = load<f32>(trig$)
        if (((trig > 0) && (prevTrig <= 0))) {
          const offsetClamped: f32 = clamp(load<f32>(offset$), 0, 1)
          angle = offsetClamped
          lastOutput = 0
        }
        prevTrig = trig
        if ((hzClamped > 0)) {
          phaseInc = (hzClamped / sampleRate)
        }
        saw = (((2 * angle) - 1) - polyBlep(angle, phaseInc))
        integrated = ((phaseInc * saw) + ((1 - phaseInc) * lastOutput))
        output = (integrated * 6)
        lastOutput = integrated
        angle = (angle + phaseInc)
        if ((angle >= 1)) {
          angle = fract01(angle)
        }
        store<f32>(output$, output)
        output$ += 4
        hz$ += 4
        offset$ += 4
        trig$ += 4
      })
    }

    this.angle = angle
    this.lastOutput = lastOutput
    this.prevTrig = prevTrig
  }
}